requirements.txt
# Основные зависимости
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
alembic==1.12.1
email-validator==2.1.0
# Test database
pytest-asyncio==0.21.0
psycopg2-binary==2.9.9
sqlalchemy-utils==0.41.1
asyncpg==0.29.0

# Testing dependencies
pytest==7.4.0
pytest-asyncio==0.21.0
pytest-cov==4.1.0
pytest-random-order==1.1.0
allure-pytest==2.13.0

# Mocking and test data
pytest-mock==3.11.1
factory-boy==3.3.0
Faker==19.0.0

# Test utilities
freezegun==1.4.0